### 2023-01-06 #5053 @editorialbot
[PRE REVIEW]: adrt: approximate discrete Radon transform for Python
**Submitting author:** @karlotness (<a href="http://orcid.org/0000-0001-8534-2648">Karl Otness</a>)
**Repository:** https://github.com/karlotness/adrt
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0.0
