### 2023-01-12 6a4ebe9
Add classifier to mark this package as type-annotated.
