### 2023-01-12 f8f0a0c
Streamline copyright notices and add SPDX license identifiers.
We still keep the full license text in the headers since it's short
enough and makes it more likely to stay attached if some of the
sources get copied. The SPDX identifiers should help any automated
license scanners.
