+++ docs/conf.py
@@ -10,7 +10,7 @@ import adrt
 
 # Project information
 project = "adrt"
-copyright = "2023 Karl Otness, Donsub Rim"
+copyright = "Karl Otness, Donsub Rim"
 author = "Karl Otness, Donsub Rim"
 version = adrt.__version__
 release = version
