+++ LICENSE.txt
@@ -1,5 +1,4 @@
-Copyright (c) 2023 Karl Otness, Donsub Rim
-All rights reserved
+Copyright Karl Otness, Donsub Rim
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are
