### 2023-01-17 #5083 @editorialbot
[REVIEW]: adrt: approximate discrete Radon transform for Python
**Submitting author:** @karlotness (<a href="http://orcid.org/0000-0001-8534-2648">Karl Otness</a>)
**Repository:** https://github.com/karlotness/adrt
**Branch with paper.md** (empty if default branch): master
**Version:** v1.0.1
