+++ docs/requirements.txt
@@ -1,3 +1,4 @@
+--only-binary :all:
 sphinx>=5.3.0,<6
 sphinx-rtd-theme>=1.1.1,<2
 numpy
