### 2023-01-23 44e679c
Streamline catch2-downloading script.
Remove cache handling from the script since we now handle that using
just GitHub actions. Use sha512 as the hash function to validate the
downloaded header.
