### 2023-01-23 cb0cdf3
Report SHA512 checksums for release artifacts.
This is in addition to the SHA256 hashes we already support. The PyPI
web interface displays SHA256 hashes so including those are useful for
comparisons between services. Some users might want to pin longer
SHA512 hashes so can report those as well.
