### 2023-01-23 defb7e4
Update flake8-bugbear version to the latest release.
