+++ tox.ini
@@ -17,7 +17,7 @@ ignore_errors = true
 deps =
   flake8>=6.0.0,<7
   pep8-naming==0.13.3
-  flake8-bugbear==22.12.6
+  flake8-bugbear==23.1.20
   black>=22.12.0,<23
 commands =
   flake8 src/adrt/ tests/ docs/ tools/ setup.py
