### 2023-01-29 fcae9c6
Migrate to ruff for linting.
It looks like several other projects are migrating, and this removes
the need to manage flake8 versions alongside plugins.
