+++ tox.ini
@@ -15,12 +15,10 @@ basepython = python3
 skip_install = true
 ignore_errors = true
 deps =
-  flake8>=6.0.0,<7
-  pep8-naming==0.13.3
-  flake8-bugbear==23.2.13
+  ruff>=0.0.246
   black>=23.1.0,<24
 commands =
-  flake8 src/adrt/ tests/ docs/ tools/ setup.py
+  ruff check -- src/adrt/ tests/ docs/ tools/ setup.py
   black --check --diff src/adrt/ tests/ docs/ tools/ setup.py
 
 [testenv:checkver]
