+++ pyproject.toml
@@ -89,3 +89,13 @@ addopts = [
   "--import-mode=importlib",
   "--capture=no",
 ]
+
+[tool.ruff]
+target-version = "py38"
+select = ["E", "F", "W", "B", "N"]
+exclude = ["docs/_build/"]
+src = ["src"]
+line-length = 88
+respect-gitignore = false
+fix = false
+update-check = false
