### 2023-01-29 c3ab00c
Ensure that linting target Python versions stay up to date.
When we update the minimal Python version for the package we also want
to update the version targeted by the linter.
