+++ tools/version_consistency.py
@@ -138,6 +138,18 @@ def find_meta_min_python(pyproject_toml):
     return find_min_version("python", ["python" + ver_constraint])
 
 
+def find_linter_min_python(pyproject_toml):
+    with open(pyproject_toml, "rb") as pyproj_file:
+        defs = tomllib.load(pyproj_file)
+    target_version = defs["tool"]["ruff"]["target-version"]
+    ver_re = re.compile(r"^py(?P<major>\d)(?P<minor>\d+)$")
+    if re_match := ver_re.match(target_version):
+        major_ver = re_match.group("major")
+        minor_ver = re_match.group("minor")
+        return f"{major_ver}.{minor_ver}"
+    raise ValueError(f"invalid linter version target '{target_version}'")
+
+
 def find_macro_min_python(py_cpp):
     min_python = find_cpp_macro_def("Py_LIMITED_API", py_cpp)
     if not min_python.startswith("0x") or len(min_python) != 10:
@@ -255,10 +267,12 @@ if __name__ == "__main__":
     # Check Python version requirements
     meta_min_python = find_meta_min_python("pyproject.toml")
     macro_limited_api = find_macro_min_python("src/adrt/adrt_cdefs_py.cpp")
+    linter_min_python = find_linter_min_python("pyproject.toml")
     print(f"Metadata min Python: {meta_min_python}")
     print(f"Limited API macro: {macro_limited_api}")
+    print(f"Linter min Python: {linter_min_python}")
     # Check consistency
-    if meta_min_python != macro_limited_api:
+    if meta_min_python != macro_limited_api or meta_min_python != linter_min_python:
         print("Python version mismatch")
         failure = True
     print("")
