### 2023-01-29 2e8564f
Extend lint rules enforced by ruff.
Check for more possible issues by enabling additional error codes in
ruff.
