+++ pyproject.toml
@@ -92,10 +92,16 @@ addopts = [
 
 [tool.ruff]
 target-version = "py38"
-select = ["E", "F", "W", "B", "N"]
+select = ["E", "F", "W", "B", "N", "RUF", "UP", "C4", "YTT", "T10", "T20", "ICN", "SIM"]
+ignore = ["UP015", "SIM108"]
 exclude = ["docs/_build/"]
 src = ["src"]
 line-length = 88
 respect-gitignore = false
 fix = false
 update-check = false
+
+[tool.ruff.per-file-ignores]
+"setup.py" = ["ICN001"]
+"tools/version_consistency.py" = ["T20"]
+"tools/download_catch2.py" = ["T20"]
