### 2023-01-29 c10f2d3
Update noqa flag to include only enabled linter codes.
We no longer run the bugbear line length check, so no need to specify
it.
