### 2023-01-29 a2e9568
Use generator and set comprehensions where applicable.
These should perform the same amount of work, but skip the call to the
built-in functions/constructors.
