### 2023-01-29 cb05949
Remove mention of flake8 in the maintenance documentation.
