+++ tox.ini
@@ -32,7 +32,7 @@ deps =
   packaging
   pip
 commands =
-  pip install --no-deps oldest-supported-numpy
+  pip install --disable-pip-version-check --no-deps oldest-supported-numpy
   {envpython} tools/version_consistency.py {posargs}
 
 [testenv:type]
