### 2023-02-13 4e8ee1d
Rename test requirements file to remove test prefix.
This should be clear from the fact that it's in the `test/` directory,
and this change keeps the name consistent with the same file for our
documentation dependencies, and with the standard name used for
requirements files more generally.
