+++ docs/install.rst
@@ -50,7 +50,7 @@ help confirm whether the routines are operating correctly. We use
 `pytest <https://pytest.org/>`__ as a test runner. Once the package is
 installed, navigate to the root directory of the repository and run::
 
-  $ python -m pip install -r tests/test_requirements.txt
+  $ python -m pip install -r tests/requirements.txt
   $ python -m pytest
 
 However, if you are trying to modify the package or make a
