+++ pyproject.toml
@@ -62,7 +62,7 @@ version = {attr = "adrt.__version__"}
 [tool.cibuildwheel]
 skip = ["*-musllinux*", "pp*"]
 build-frontend = "build"
-before-test = "pip install -r {package}/tests/test_requirements.txt"
+before-test = "pip install -r {package}/tests/requirements.txt"
 test-command = "pytest {package}/tests"
 test-skip = "*-macosx_universal2:arm64 cp3{10,11,12}-win32 cp312-*"
 
