+++ tox.ini
@@ -5,7 +5,7 @@ isolated_build = true
 [testenv]
 description = Run tests
 deps =
- -rtests/test_requirements.txt
+ -rtests/requirements.txt
 commands =
   pytest {posargs}
 
