+++ .github/workflows/test.yml
@@ -70,7 +70,7 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install --upgrade build
-        python -m pip install --upgrade -r tests/test_requirements.txt
+        python -m pip install --upgrade -r tests/requirements.txt
     - name: Build package
       run: python -m build --wheel
       env:
