### 2023-02-14 24124a1
Add link to download data file used in CT example.
This is the same file as used in the sinogram demonstration, but we
should provide a link here as well.
