### 2023-02-14 a2bf69e
Rename "match" variables, adding "re" prefix.
"match" is a soft keyword in Python 3.10. This doesn't cause any
problems with our use, and may not change in the future, but still we
can avoid any possible issues entirely by just renaming the variables.
