### 2023-02-14 071ce1e
Remove leading and trailing spaces from regex match before parsing.
This should fix issues with Python 3.9's ast parsing which interpreted
leading spaces as invalid indentation. See issue #3 for more details.
