+++ tox.ini
@@ -17,8 +17,8 @@ ignore_errors = true
 deps =
   flake8>=6.0.0,<7
   pep8-naming==0.13.3
-  flake8-bugbear==23.1.20
-  black>=22.12.0,<23
+  flake8-bugbear==23.2.13
+  black>=23.1.0,<24
 commands =
   flake8 src/adrt/ tests/ docs/ tools/ setup.py
   black --check --diff src/adrt/ tests/ docs/ tools/ setup.py
@@ -39,7 +39,7 @@ commands =
 description = Check types in library source
 basepython = python3
 deps =
-  mypy>=0.991
+  mypy>=1.0.0
 commands =
   mypy --strict --package adrt
 
