### 2023-02-16 6d9205b
Enable ruff NumPy-specific lint rules.
