+++ pyproject.toml
@@ -92,7 +92,7 @@ addopts = [
 
 [tool.ruff]
 target-version = "py38"
-select = ["E", "F", "W", "B", "N", "RUF", "UP", "C4", "YTT", "T10", "T20", "ICN", "SIM"]
+select = ["E", "F", "W", "B", "N", "RUF", "UP", "C4", "YTT", "T10", "T20", "ICN", "SIM", "NPY"]
 ignore = ["UP015", "SIM108"]
 exclude = ["docs/_build/"]
 src = ["src"]
