+++ tox.ini
@@ -15,7 +15,7 @@ basepython = python3
 skip_install = true
 ignore_errors = true
 deps =
-  ruff>=0.0.246
+  ruff>=0.0.247
   black>=23.1.0,<24
 commands =
   ruff check -- src/adrt/ tests/ docs/ tools/ setup.py
