### 2023-02-27 4c8f0cd
Alter computation of unaligned array offsets.
Always offset by one byte since the previous calculation led to
unexpectedly *aligned* arrays. A one byte offset should more reliably
misalign the buffers for this test.
