### 2023-02-27 91e969d
Allow ValueError for negative integers.
CPython produces an OverflowError error here while PyPy produces
ValueError. For our purposes, we can accept both.
