### 2023-03-01 3082805
Update documentation build requirements.
Update limited ranges for Sphinx and the theme to the latest versions.
Drop the limits on matplotlib version range since we should generally
be getting the latest version anyway.
