+++ .readthedocs.yml
@@ -4,7 +4,7 @@ sphinx:
 build:
   os: ubuntu-22.04
   tools:
-    python: '3.10'
+    python: '3.11'
   apt_packages:
     - g++
     - imagemagick
