### 2023-03-05 69fa21b
Remove link to license file in readme.
These links only work on GitHub and are broken elsewhere (in
particular on PyPI).
