+++ README.md
@@ -37,8 +37,8 @@ This implementation is based on descriptions in several publications:
 
 ## License
 
-The code in this repository is licensed under a 3-clause BSD license.
-See [LICENSE.txt](LICENSE.txt) for the license text.
+This software is distributed under the 3-clause BSD license. See
+LICENSE.txt for the license text.
 
 We also make available several pre-built binary copies of this
 software. The binary build for Windows includes additional license
