### 2023-03-13 karlotness/adrt#4 @fruzsinaagocs
Time complexity
Hi,

I checked the time complexity of ADRT for NxN images similar to those in the quickstart example:

``` python
import numpy as np
import time
import adrt

nmin = 2
nmax = 14
num = nmax - nmin + 1
ns = np.logspace(nmin, nmax, num = num, base = 2)
times = np.zeros(num)
timen = 10000
for i, n in enumerate(ns):
    nint = int(n)
    xs = np.linspace(-1, 1, nint)
    x, y = np.meshgrid(xs, xs)
    img = 0.5 * ((np.abs(x - 0.25) + np.abs(y)) < 0.7).astype(np.float32)
    img[:, np.random.randint(0, nint-1)] = 1
    img[np.random.randint(0, nint-1), :] = 1
    start = time.time()
    if n > 100:
        timen = 1000
    if n > 1000:
        timen = 10
    if n > 10000:
        timen = 1
    for j in range(timen):
       result = adrt.adrt(img)
    end = time.time()
    times[i] = (end - start)/timen
```

This gave me:
[scaling.pdf](https://github.com/karlotness/adrt/files/10962322/scaling.pdf)

So the claim that the algorithm is O(N^2 log N) is clearly true (in the limit of large N), but I was wondering if you also noted that between N = 2^7 and N = 2^10 (which may be typical image sizes) it scales more like N^3?

Thanks!
