### 2023-03-13 #5083 @fruzsinaagocs
Hi @diehlpk, @karlotness, I finished my preliminary review, and everything looks good. I submitted karlotness/adrt#4 which investigates the time complexity of ADRT, found that the complexity claim was true with something mildly interesting happening at intermediate image sizes, which I think might be worth commenting on.
