### 2023-03-14 karlotness/adrt#4 @karlotness
Thanks for running those tests; running your timing code I do see the same thing.

I suspect the bend in the curve is largely driven by cache effects. The time for the adrt to run is bounded mainly by memory bandwidth so the time required to pass through both of the internal buffers is what should mainly drive the scaling. It really should be scaling as $O(n^2 \log_2(n))$ since that's the number of memory accesses, but with different memory access costs as the buffers grow.

In that plot, I think what's mostly happening is that as the size of the image grows larger, eventually it becomes too large for each layer in your processor's cache, finally requiring access to main memory. That's the point where the scaling should become really consistent.

To try to look at that, I reran some measurements similar to the timings you ran, but adjusting the batch dimension so that the buffers are closer to the same size in each trial. So we start with a very large batch for the 2x2 images, then divide by a factor of 4 each time the image dimension doubles.

  <summary>Test Code</summary>

```python
import numpy as np
import time
import statistics
import adrt
import math
from tqdm.auto import tqdm

TRIALS = 3
ns = [2**i for i in range(1, 15)]
batch_sizes = [(max(ns) // n)**2 for n in ns]
times = []
buf_sizes = []

rng = np.random.default_rng()

for n, batch_size in tqdm(zip(ns, batch_sizes), total=len(ns)):
    # Init image
    img = rng.normal(size=(batch_size, n, n))
    # Run trials
    iter_times = []
    for i in range(TRIALS):
        start = time.perf_counter()
        res = adrt.adrt(img)
        end = time.perf_counter()
        buf_size = res[0].nbytes
        res = None
        iter_times.append((end - start) / batch_size)
    buf_sizes.append(buf_size)
    times.append(statistics.mean(iter_times))
```

With that I get a more consistent trend, without the odd bend in the timing curve early on (plotted below with a trend line).

![fig](https://user-images.githubusercontent.com/8847665/224882839-f3e6cd41-1576-459d-bf61-8591112640c2.png)

Does that seem to make sense?

Thanks again for taking a look
