### 2023-03-14 #5083 @fruzsinaagocs
@diehlpk, The issue has now been addressed (in short, @karlotness demonstrated that the interesting 'dip' in the complexity curve was driven by cache effects), I'm happy to accept the paper for publication. Thank you for asking me to review!
