### 2023-03-14 #5083 @diehlpk
@karlotness Would it be much work to add your scaling plot from here (https://github.com/karlotness/adrt/issues/4) to the paper?
