+++ joss/paper.md
@@ -20,7 +20,7 @@ affiliations:
    index: 1
  - name: Washington University in St. Louis, USA
    index: 2
-date: 28 December 2022
+date: 14 March 2023
 bibliography: paper.bib
 ---
 
