### 2023-03-15 269dabe
Insert figure with timing results.
