### 2023-03-15 #5083 @diehlpk
@karlotness If you reference the figure in the text and provided on which CPU you executed the code, I am happy to accept the paper.
