### 2023-03-15 5a7defe
Add information on CPU used for timing tests.
