### 2023-03-15 #5083 @karlotness
Thanks, that's a good point. I've just added that to the figure caption.

The figure also should be referenced in the text, in the discussion of the time complexity of the ADRT.
