+++ src/adrt/__init__.py
@@ -61,7 +61,7 @@ __all__: typing.Final[typing.Sequence[str]] = [
     "utils",
     "core",
 ]
-__version__: typing.Final[str] = "1.0.1.dev"
+__version__: typing.Final[str] = "1.0.1"
 
 
 _F = typing.TypeVar("_F", np.float32, np.float64)
