### 2023-03-15 2a115a4
Extend timeout for release binary builds.
The Windows runs in particular are sometimes very slow, so we should
allow more time.
