+++ .github/workflows/release.yml
@@ -114,7 +114,7 @@ jobs:
         if-no-files-found: error
   wheel:
     runs-on: ${{ matrix.os }}
-    timeout-minutes: 30
+    timeout-minutes: 60
     needs: sdist
     strategy:
       matrix:
