### 2023-03-15 #5083 @karlotness
Should be set on both of those, I think.

I've added a new release version `1.0.1` with a Git tag of `v1.0.1`

I added the repository to Zenodo.
The DOI for this version is [10.5281/zenodo.7738254](https://doi.org/10.5281/zenodo.7738254)
and the DOI for "all versions" there is [10.5281/zenodo.7738253](https://doi.org/10.5281/zenodo.7738253)

Thanks!
