+++ joss/paper.bib
@@ -42,7 +42,7 @@
 
 @article{Gotz96fdrt,
   author = {W.A. G\"otz and H.J. Druckm\"uller},
-  title = {A fast digital radon transform: an efficient means for evaluating the hough transform},
+  title = {A fast digital {R}adon transform: an efficient means for evaluating the {H}ough transform},
   journal = {Pattern Recognition},
   volume = 29,
   number = 4,
@@ -54,7 +54,7 @@
 
 @article{ACDISS2008,
   author = {Averbuch, A. and Coifman, R. R. and Donoho, D. L. and Israeli, M. and Shkolnisky, Y. and Sedelnikov, I.},
-  title = {A Framework for Discrete Integral Transformations II—The 2D Discrete Radon Transform},
+  title = {A Framework for Discrete Integral Transformations II—The 2D Discrete {R}adon Transform},
   journal = {SIAM Journal on Scientific Computing},
   volume = 30,
   number = 2,
@@ -66,7 +66,7 @@
 @article{Beylkin1987,
   author = {G. Beylkin},
   journal = {IEEE Transactions on Acoustics, Speech, and Signal Processing},
-  title = {Discrete radon transform},
+  title = {Discrete {R}adon transform},
   year = 1987,
   volume = 35,
   number = 2,
@@ -76,7 +76,7 @@
 
 @article{GreengardLee2004,
   author = {Greengard, Leslie and Lee, June-Yub},
-  title = {Accelerating the Nonuniform Fast Fourier Transform},
+  title = {Accelerating the Nonuniform Fast {F}ourier Transform},
   journal = {SIAM Review},
   volume = 46,
   number = 3,
@@ -87,7 +87,7 @@
 
 @article{BMK2019,
   author = {Barnett, Alexander H. and Magland, Jeremy and {af Klinteberg}, Ludvig},
-  title = {A Parallel Nonuniform Fast Fourier Transform Library Based on an ``Exponential of Semicircle'' Kernel},
+  title = {A Parallel Nonuniform Fast {F}ourier Transform Library Based on an ``Exponential of Semicircle'' Kernel},
   journal = {SIAM Journal on Scientific Computing},
   volume = 41,
   number = 5,
@@ -99,7 +99,7 @@
 
 @article{Rim2018,
   author = {Rim, Donsub},
-  title = {Dimensional Splitting of Hyperbolic Partial Differential Equations Using the Radon Transform},
+  title = {Dimensional Splitting of Hyperbolic Partial Differential Equations Using the {R}adon Transform},
   journal = {SIAM Journal on Scientific Computing},
   volume = 40,
   number = 6,
@@ -160,7 +160,7 @@
 @inproceedings{radonopencl,
   author={B{\"u}cker, H. Martin and Seidler, Ralf and Neuh{\"a}user, David and Beier, Tobias},
   booktitle={2015 IEEE 9th International Symposium on Embedded Multicore/Many-core Systems-on-Chip},
-  title={The Approximate Discrete Radon Transform: A Case Study in Auto-Tuning of {OpenCL} Implementations},
+  title={The Approximate Discrete {R}adon Transform: A Case Study in Auto-Tuning of {OpenCL} Implementations},
   year=2015,
   pages={219-226},
   doi={10.1109/MCSoC.2015.38},
