### 2023-03-16 #5083 @danielskatz
👋 @karlotness - I'm the track editor who will handle the final processing for this submission. While proofreading it, I found a number of small issues, as indicated in https://github.com/karlotness/adrt/pull/5 - Please merge this, or let me know what you disagree with, then we can continue the process of acceptance and publication.
